requirements.txt
requests==2.13.0
tornado==4.4.2
Pillow==4.0.0
beautifulsoup4==4.5.3
lxml==3.7.3
